#26: Move code into sub package db
