+++ balta/src/main/scala/za/co/absa/db/balta/classes/DBConnection.scala
@@ -14,21 +14,18 @@
  * limitations under the License.
  */
 
-import Dependencies._
+package za.co.absa.db.balta.classes
 
-lazy val scala211 = "2.11.12"
-lazy val scala212 = "2.12.18"
-lazy val scala213 = "2.13.11"
+import za.co.absa.db.balta.classes.simple.ConnectionInfo
 
-lazy val supportedScalaVersions: Seq[String] = Seq(scala211, scala212 , scala213)
+import java.sql.{Connection, DriverManager}
 
-ThisBuild / scalaVersion := scala212
+class DBConnection(val connection: Connection) extends AnyVal
 
-ThisBuild / versionScheme := Some("early-semver")
-
-lazy val balta = (project in file("balta"))
-  .settings(
-    name := "balta",
-    crossScalaVersions := supportedScalaVersions,
-    libraryDependencies ++= libDependencies
-  )
+object DBConnection {
+  def apply(connectionInfo: =>ConnectionInfo): DBConnection = {
+    val connection = DriverManager.getConnection(connectionInfo.dbUrl, connectionInfo.username, connectionInfo.password)
+    connection.setAutoCommit(false)
+    new DBConnection(connection)
+  }
+}
