+++ balta/src/main/scala/za/co/absa/db/balta/classes/DBTable.scala
@@ -14,10 +14,12 @@
  * limitations under the License.
  */
 
-package za.co.absa.balta.classes
+package za.co.absa.db.balta.classes
 
-import za.co.absa.balta.classes.setter.{AllowedParamTypes, Params, SetterFnc}
-import za.co.absa.balta.classes.setter.Params.NamedParams
+import za.co.absa.db.balta.classes.setter.{AllowedParamTypes, Params, SetterFnc}
+import za.co.absa.db.balta.classes.setter.Params.NamedParams
+import za.co.absa.db.balta.classes.setter.Params.NamedParams
+import za.co.absa.db.balta.classes.setter.{AllowedParamTypes, Params}
 
 /**
   * This class represents a database table. It allows to perform INSERT, SELECT and COUNT operations on the table easily.
