+++ balta/src/main/scala/za/co/absa/db/balta/classes/setter/CustomDBType.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package za.co.absa.balta.classes.setter
+package za.co.absa.db.balta.classes.setter
 
 /**
  * This is a case class representing a custom DB type.
