+++ balta/src/main/scala/za/co/absa/db/balta/classes/setter/Params.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package za.co.absa.balta.classes.setter
+package za.co.absa.db.balta.classes.setter
 
 import scala.collection.immutable.ListMap
