+++ balta/src/main/scala/za/co/absa/db/balta/classes/setter/SetterFnc.scala
@@ -14,13 +14,12 @@
  * limitations under the License.
  */
 
-package za.co.absa.balta.classes.setter
-
-import za.co.absa.balta.classes.JsonBString
+package za.co.absa.db.balta.classes.setter
 
 import java.sql.{Date, PreparedStatement, Time, Timestamp, Types => SqlTypes}
 import java.util.UUID
 import org.postgresql.util.PGobject
+import za.co.absa.db.balta.classes.simple.JsonBString
 
 import java.time.{Instant, LocalDate, LocalTime, OffsetDateTime, ZoneId, ZoneOffset}
