+++ balta/src/main/scala/za/co/absa/db/balta/classes/simple/JsonBString.scala
@@ -14,18 +14,6 @@
  * limitations under the License.
  */
 
-package za.co.absa.balta
+package za.co.absa.db.balta.classes.simple
 
-import za.co.absa.balta.classes.DBConnection
-
-import java.sql.Connection
-import scala.language.implicitConversions
-
-package object implicits {
-
-  /**
-   * This implicit conversion allows to use a DBConnection at any place where as a JDBC Connection is required.
-   */
-  implicit def dbConnectionToJdbcConnection(in: DBConnection): Connection = in.connection
-
-}
+case class JsonBString(value: String) extends AnyVal
